refactor(BA-7054): add operation_type and spec to the pure-ABC action bases
The pure-ABC action bases dropped operation_type()/spec(), but the reporter
subscription key ("entity:operation" from spec().type()) and audit-log records
depend on them. Add operation_type() and spec() to each of the three bases
independently (no shared supertype), retype ActionSpec.entity_type to the new
open EntityType, and wrap the legacy BaseAction.spec() value accordingly.
