+++ src/ai/backend/manager/actions/scope/base.py
@@ -3,6 +3,7 @@ from collections.abc import Sequence
 
 from ai.backend.common.data.entity.types import EntityType, ScopeRef
 from ai.backend.common.data.permission.types import Permission
+from ai.backend.manager.actions.types import ActionOperationType, ActionSpec
 
 
 class BaseScopeAction(ABC):
@@ -21,6 +22,20 @@ class BaseScopeAction(ABC):
 
     @classmethod
     @abstractmethod
+    def operation_type(cls) -> ActionOperationType:
+        """Return the operation that this action performs within the scopes."""
+        raise NotImplementedError
+
+    @classmethod
+    @abstractmethod
     def required_permission(cls) -> Permission:
         """Return the permission required to perform this action."""
         raise NotImplementedError
+
+    @classmethod
+    def spec(cls) -> ActionSpec:
+        """Return the "entity:operation" spec keying reporter subscriptions and audit records."""
+        return ActionSpec(
+            entity_type=cls.entity_type(),
+            operation_type=cls.operation_type(),
+        )
