+++ src/ai/backend/manager/actions/single_entity/base.py
@@ -3,6 +3,7 @@ from abc import ABC, abstractmethod
 from ai.backend.common.data.entity.types import EntityType
 from ai.backend.common.data.permission.types import Permission
 from ai.backend.common.identifier.entity import EntityID
+from ai.backend.manager.actions.types import ActionOperationType, ActionSpec
 
 
 class BaseSingleEntityAction(ABC):
@@ -12,6 +13,12 @@ class BaseSingleEntityAction(ABC):
     @abstractmethod
     def entity_type(cls) -> EntityType:
         """Return the type of entity that this action applies to."""
+        raise NotImplementedError
+
+    @classmethod
+    @abstractmethod
+    def operation_type(cls) -> ActionOperationType:
+        """Return the operation that this action performs on the entity."""
         raise NotImplementedError
 
     @abstractmethod
@@ -24,3 +31,11 @@ class BaseSingleEntityAction(ABC):
     def required_permission(cls) -> Permission:
         """Return the permission required to perform this action."""
         raise NotImplementedError
+
+    @classmethod
+    def spec(cls) -> ActionSpec:
+        """Return the "entity:operation" spec keying reporter subscriptions and audit records."""
+        return ActionSpec(
+            entity_type=cls.entity_type(),
+            operation_type=cls.operation_type(),
+        )
